useCloudAnimation hook
