+++ packages/design-system/styles/globals.css
@@ -279,3 +279,60 @@
     --gradient-angle: 360deg;
   }
 }
+
+.cloud-container {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-color: #e0f0ff;
+  overflow: hidden;
+}
+
+.cloud-layer {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;
+  transition: transform 2s ease-out;
+  will-change: transform;
+}
+
+#cloud1 {
+  background-image: url('/clouds/back_1-min.png');
+  opacity: 0.5;
+  z-index: 1;
+}
+
+#cloud2 {
+  background-image: url('/clouds/back_2-min.png');
+  opacity: 0.5;
+  z-index: 2;
+}
+
+#cloud3 {
+  background-image: url('/clouds/back_3-min.png');
+  opacity: 0.5;
+  z-index: 4;
+}
+
+#cloud4 {
+  background-image: url('/clouds/back_1_proc-min.png');
+  opacity: 0.3;
+  z-index: 2;
+}
+
+#cloud5 {
+  background-image: url('/clouds/back_4_proc-min.png');
+  opacity: 0.3;
+  z-index: 3;
+}
+
+#cloud6 {
+  background-image: url('/clouds/back_3_proc-min.png');
+  opacity: 0.3;
+  z-index: 1;
+}
